Implement proposal changes from PR comments. Move to exercise_solutions folder
